requirements.txt
flask
flask_jwt_extended
bcrypt
flask_session
cs50
datetime
cryptography
werkzeug
tweepy
python-dotenv
utils
